Add support for molly-guard
Update the script to always call the "unguarded" versions of reboot and
shutdown, if molly-guard is installed.

Molly-guard is a package that replaces the reboot and shutdown commands
with "guarded" versions that prompt you before executing the command,
to avoid accidentally rebooting the wrong system.  This is a common problem
if you ssh into other systems and often reboot those systems.

Normally, molly-guard only triggers on an ssh logins, but it can be
configured to always trigger.  In this case, grub-reboot-picker triggers
the guard when it tries to reboot or shut down, and that interrupts the
actual reboot process.