Use grub-mkconfig to read grub entries. Don't require pkexec to run grub-reboot-picker script. This allows the tray icon to have a transparent background.
For some reason creating an app indicator in a script that's running as sudo causes it to have a black background on hover.
By using polkit to allow running grub-mkconfig, the main grub-reboot-picker script can run as normal user.
And running as normal user allows cleaner icon in the tray.